feat(amount): Add Amount struct to replace f64 and to manage asset amount held in account
Addressing the elephant in the room: Rounding error with f64, by adding an Amount class managing the asset amount held in account. The precision is four places past the decimal, it is memory efficient by leveraging an i64 behind the hood. It has however limitations in max and min values acceptable, but the overflow and underflow are checked.